Move BioSamplesDependentTest class from subs-test-categories
