+++ build.gradle
@@ -57,7 +57,6 @@ dependencies {
 
     testCompile("junit:junit")
     testCompile("org.springframework.boot:spring-boot-starter-test")
-    testCompile("uk.ac.ebi.subs:subs-test-categories:0.1.0-SNAPSHOT")
 }
 
 task sourceJar(type: Jar) { from sourceSets.main.allJava }
