+++ src/test/java/uk/ac/ebi/subs/agent/utils/BioSamplesDependentTest.java
@@ -1,0 +1,7 @@
+package uk.ac.ebi.subs.agent.utils;
+
+/**
+ * Marker interface for tests that rely on BioSamples DB
+ */
+public interface BioSamplesDependentTest {
+}
